requirements.txt
asgiref==3.7.2
Django==5.0.3
django-filter==24.1
djangorestframework==3.14.0
install==1.3.5
Markdown==3.6
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
